chore: package 구조 변경
- AS-IS : step2.lotto
- TO-BE : lotto
- link 깨짐 수정
