+++ src/main/kotlin/lotto/PlayLotto.kt
@@ -1,0 +1,8 @@
+package lotto
+
+import lotto.controller.LottoController
+import lotto.service.RandomLottoGenerator
+
+fun main() {
+    LottoController(RandomLottoGenerator()).playLotto()
+}
