+++ src/main/kotlin/lotto/domain/Lottos.kt
@@ -1,0 +1,18 @@
+package lotto.domain
+
+class Lottos private constructor(val elements: List<Lotto>) {
+    fun match(winningNumber: WinningNumber): PlayResults {
+        val playResults = PlayResults()
+        elements.map {
+            val matchResult = it.match(winningNumber)
+            playResults.add(it, matchResult)
+        }
+        return playResults
+    }
+
+    companion object {
+        fun of(lottos: List<Lotto>): Lottos {
+            return Lottos(lottos)
+        }
+    }
+}
