+++ src/main/kotlin/lotto/domain/PlayResult.kt
@@ -1,0 +1,7 @@
+package lotto.domain
+
+class PlayResult private constructor(val lotto: Lotto, matchResult: MatchResult) {
+    companion object {
+        fun of(lotto: Lotto, matchResult: MatchResult): PlayResult = PlayResult(lotto, matchResult)
+    }
+}
