+++ src/main/kotlin/lotto/domain/PurchaseItem.kt
@@ -1,0 +1,11 @@
+package lotto.domain
+
+class PurchaseItem private constructor(buyAmount: BuyAmount, val lottos: Lottos) {
+    val buyAmount: Int = buyAmount.value
+    val tryCount: Int = buyAmount.tryCount
+
+    companion object {
+        fun of(buyAmount: BuyAmount, lottos: Lottos): PurchaseItem =
+            PurchaseItem(buyAmount, lottos)
+    }
+}
