--- src/main/kotlin/step2/lotto/domain/PlayResult.kt
@@ -1,7 +1,0 @@
-package step2.lotto.domain
-
-class PlayResult private constructor(val lotto: Lotto, matchResult: MatchResult) {
-    companion object {
-        fun of(lotto: Lotto, matchResult: MatchResult): PlayResult = PlayResult(lotto, matchResult)
-    }
-}
